+++ src/vscode/notebook/_types.ts
@@ -2,5 +2,6 @@ export type QueryLanguage =
 	"DAX"
 |	"API"
 |	"CMD"
+|	"TMSL"
 ;